refactor: enhancing payment management with improved filtering, sorting, and pagination features #17
