+++ bootstrap/providers.php
@@ -2,5 +2,4 @@
 
 return [
     App\Providers\AppServiceProvider::class,
-    App\Providers\VoltServiceProvider::class,
 ];
