+++ composer.json
@@ -15,7 +15,6 @@
         "laravel/tinker": "^2.10.1",
         "livewire/flux": "^2.0",
         "livewire/livewire": "^3.6",
-        "livewire/volt": "^1.6.7",
         "spatie/laravel-permission": "^6.16"
     },
     "require-dev": {
