+++ resources/lang/en/khairat.php
@@ -83,6 +83,8 @@ return [
     'all_months' => 'All Months',
     'reset_filters' => 'Reset Filters',
     'no_payments_found' => 'No payments found.',
+    'you_have_not_made_any_payments' => 'You have not made any payments yet.',
+    'make_a_payment' => 'Make a Payment',
     'date' => 'Date',
     'reference' => 'Reference',
     'amount' => 'Amount',
@@ -94,6 +96,16 @@ return [
     'view' => 'View',
     'search' => 'Search',
     'reference_or_notes' => 'Reference or Notes',
+    'advanced_view' => 'Advanced View',
+    'simple_view' => 'Simple View',
+    'filter' => 'Filter',
+    'upload_receipt' => 'Upload Receipt',
+    'receipt_image' => 'Receipt Image',
+    'notes_optional' => 'Notes (Optional)',
+    'upload' => 'Upload',
+    'uploading' => 'Uploading...',
+    'upload_payment_receipt' => 'Upload Payment Receipt',
+    'accepted_formats' => 'Accepted formats: JPG, PNG, PDF. Max size: 2MB',
 
     // Receipts
     'receipt' => 'Receipt',
