+++ resources/lang/ms/khairat.php
@@ -83,6 +83,8 @@ return [
     'all_months' => 'Semua Bulan',
     'reset_filters' => 'Set Semula Penapis',
     'no_payments_found' => 'Tiada pembayaran dijumpai.',
+    'you_have_not_made_any_payments' => 'Anda belum membuat sebarang pembayaran.',
+    'make_a_payment' => 'Buat Pembayaran',
     'date' => 'Tarikh',
     'reference' => 'Rujukan',
     'amount' => 'Jumlah',
@@ -94,6 +96,16 @@ return [
     'view' => 'Lihat',
     'search' => 'Cari',
     'reference_or_notes' => 'Rujukan atau Nota',
+    'advanced_view' => 'Paparan Lanjutan',
+    'simple_view' => 'Paparan Ringkas',
+    'filter' => 'Tapis',
+    'upload_receipt' => 'Muat Naik Resit',
+    'receipt_image' => 'Imej Resit',
+    'notes_optional' => 'Nota (Pilihan)',
+    'upload' => 'Muat Naik',
+    'uploading' => 'Sedang Muat Naik...',
+    'upload_payment_receipt' => 'Muat Naik Resit Pembayaran',
+    'accepted_formats' => 'Format yang diterima: JPG, PNG, PDF. Saiz maksimum: 2MB',
 
     // Receipts
     'receipt' => 'Resit',
